refactor: Removal of deprecated code
- Removed deprecated methods, properties, classes, modules and data
file.
-
Refactored Validators.py to validators.py to follow PEP8
guidelines.
- Added example JSON for impute values.
