+++ CAPICE_example/example_impute_values.json
@@ -1,0 +1,1 @@
+{"Consequence": "UNKNOWN", "motifEHIPos": 0, "PolyPhenCat": "unknown", "PolyPhenVal": 0.0, "motifECount": 0, "cDNApos": 0.0, "relcDNApos": 0.0, "SIFTcat": "UD", "SIFTval": 0.0, "protPos": 0.0, "relProtPos": 0.0, "Domain": "UD", "oAA": "unknown", "nAA": "unknown", "motifEScoreChng": 0.0, "CDSpos": 0.0, "relCDSpos": 0.0, "ref": "N", "alt": "N"}
