--- CAPICE_example/not_so_default.cfg
@@ -1,36 +1,0 @@
-[DEFAULTS]
-# Set a genome build if not present in file header in integer. Can be left "missing".
-genomebuild = 20
-
-# Set a VEP version if not present in file header in float. Can be left "missing".
-vepversion = missing
-
-[OVERWRITES]
-# Set the string to overwrite what impute file should be used (found as the name in the init of a imputing file). Can be left "missing".
-imputefile = missing
-
-# Set the string to overwrite what model file should be used (found as the name in the init of a model file). Can be left "missing".
-modelfile = missing
-
-[DATAFILES]
-# The location of the reference genome.
-reference = missing
-
-[TRAINING]
-# Make the input data balanced. 0 = no balancing, 1 = balancing
-makebalanced = 0
-
-# Load the default hyper parameters used by Li et al. 0 = no, 1 = yes.
-default = 0
-
-# The location of a JSON containing specific hyper parameters to be used in training. Can be left "missing".
-specifieddefaults = missing
-
-# Float defining the size of the initial split of input data into a validation and train/test dataset. Range = 0.0-1.0
-split = 0.0
-
-# Float defining the size of the internal size of the testing dataset. Range = 0.1-1.0
-traintestsize = 0.2
-
-# Use the early exit feature of the training script. 0 = no early exit, 1 = early exit.
-earlyexit = 0
