--- CAPICE_example/specified_defaults.json
@@ -1,1 +1,0 @@
-{"learning_rate": 0.5, "n_estimators": 10, "max_depth": 10}
