+++ src/main/python/core/global_manager.py
@@ -11,127 +11,9 @@ class CapiceManager:
     class __CapiceManager:
         def __init__(self):
             self.property_checker = PropertyChecker()
-            self._now = datetime.now()
-            self.overwrite_impute = None
-            self.overwrite_model = None
-            self.grch_build = None
-            self.config_grch_build = None
-            self.force = False
             self.loglevel = None
             self.critical_logging_only = False
-            self.annotation_features = []
             self.output_filename = ''
-            self.reference_genome = None
-            self.vep_version = None
-            self.config_vep_version = None
-            self.config_loc = None
-
-        @property
-        def now(self):
-            """
-            Deprecated
-            """
-            return self._now
-
-        @property
-        def overwrite_impute(self):
-            """
-            Deprecated
-            """
-            return self._overwrite_impute
-
-        @overwrite_impute.setter
-        def overwrite_impute(self, value=None):
-            """
-            Deprecated
-            """
-            self.property_checker.check_property(value=value,
-                                                 expected_type=str,
-                                                 include_none=True)
-            self._overwrite_impute = value
-
-        @property
-        def overwrite_model(self):
-            """
-            Deprecated
-            """
-            return self._overwrite_model
-
-        @overwrite_model.setter
-        def overwrite_model(self, value):
-            """
-            Deprecated
-            """
-            self.property_checker.check_property(value=value,
-                                                 expected_type=str,
-                                                 include_none=True)
-            self._overwrite_model = value
-
-        @property
-        def grch_build(self):
-            """
-            Deprecated
-            """
-            return self._grch_build
-
-        @grch_build.setter
-        def grch_build(self, value):
-            """
-            Deprecated
-            """
-            self.property_checker.check_property(value=value,
-                                                 expected_type=int,
-                                                 include_none=True)
-            self._grch_build = value
-
-        @property
-        def config_grch_build(self):
-            """
-            Deprecated
-            """
-            return self._config_grch_build
-
-        @config_grch_build.setter
-        def config_grch_build(self, value):
-            """
-            Deprecated
-            """
-            self.property_checker.check_property(value=value,
-                                                 expected_type=int,
-                                                 include_none=True)
-            self._config_grch_build = value
-
-        @property
-        def annotation_features(self):
-            """
-            Deprecated
-            """
-            return self._annotation_features
-
-        @annotation_features.setter
-        def annotation_features(self, value):
-            """
-            Deprecated
-            """
-            self.property_checker.check_property(value=value,
-                                                 expected_type=list)
-            self._annotation_features = value
-
-        @property
-        def force(self):
-            """
-            Deprecated
-            """
-            return self._force
-
-        @force.setter
-        def force(self, value):
-            """
-            Deprecated
-            """
-            self.property_checker.check_property(value=value,
-                                                 expected_type=bool)
-            self._force = value
 
         @property
         def loglevel(self):
@@ -163,172 +45,7 @@ class CapiceManager:
             self.property_checker.check_property(value=value, expected_type=str)
             self._output_filename = value
 
-        @property
-        def reference_genome(self):
-            """
-            Deprecated
-            """
-            return self._reference_genome
-
-        @reference_genome.setter
-        def reference_genome(self, value):
-            """
-            Deprecated
-            """
-            self.property_checker.check_property(value=value,
-                                                 expected_type=str,
-                                                 include_none=True)
-            self._reference_genome = value
-
-        @property
-        def vep_version(self):
-            """
-            Deprecated
-            """
-            return self._vep_version
-
-        @vep_version.setter
-        def vep_version(self, value):
-            """
-            Deprecated
-            """
-            self.property_checker.check_property(value=value,
-                                                 expected_type=float,
-                                                 include_none=True)
-            self._vep_version = value
-
-        @property
-        def config_vep_version(self):
-            """
-            Deprecated
-            """
-            return self._config_vep_version
-
-        @config_vep_version.setter
-        def config_vep_version(self, value):
-            """
-            Deprecated
-            """
-            self.property_checker.check_property(value=value,
-                                                 expected_type=float,
-                                                 include_none=True)
-            self._config_vep_version = value
-
-        @property
-        def config_loc(self):
-            """
-            Deprecated
-            """
-            return self._config_loc
-
-        @config_loc.setter
-        def config_loc(self, value):
-            """
-            Deprecated
-            """
-            if value is not None and not os.path.isfile(value):
-                raise FileNotFoundError('Given config location is not a file!')
-            self._config_loc = value
-
     instance = None
-
-    @property
-    def now(self):
-        """
-        Deprecated
-        """
-        return self._now
-
-    @now.setter
-    def now(self, value):
-        """
-        Deprecated
-        """
-        pass
-
-    @property
-    def overwrite_impute(self):
-        """
-        Deprecated
-        """
-        return self._overwrite_impute
-
-    @overwrite_impute.setter
-    def overwrite_impute(self, value=None):
-        """
-        Deprecated
-        """
-        pass
-
-    @property
-    def overwrite_model(self):
-        """
-        Deprecated
-        """
-        return self._overwrite_model
-
-    @overwrite_model.setter
-    def overwrite_model(self, value):
-        """
-        Deprecated
-        """
-        pass
-
-    @property
-    def grch_build(self):
-        """
-        Deprecated
-        """
-        return self._grch_build
-
-    @grch_build.setter
-    def grch_build(self, value):
-        """
-        Deprecated
-        """
-        pass
-
-    @property
-    def config_grch_build(self):
-        """
-        Deprecated
-        """
-        return self._config_vep_version
-
-    @config_grch_build.setter
-    def config_grch_build(self, value):
-        """
-        Deprecated
-        """
-        pass
-
-    @property
-    def annotation_features(self):
-        """
-        Deprecated
-        """
-        return self._annotation_features
-
-    @annotation_features.setter
-    def annotation_features(self, value):
-        """
-        Deprecated
-        """
-        pass
-
-    @property
-    def force(self):
-        """
-        Deprecated
-        """
-        return self._force
-
-    @force.setter
-    def force(self, value):
-        """
-        Deprecated
-        """
-        pass
 
     @property
     def loglevel(self):
@@ -388,62 +105,6 @@ class CapiceManager:
         to set the output file name that CAPICE prediction will produce.
 
         :param value: path-like
-        """
-        pass
-
-    @property
-    def reference_genome(self):
-        """
-        Deprecated
-        """
-        return self._reference_genome
-
-    @reference_genome.setter
-    def reference_genome(self, value):
-        """
-        Deprecated
-        """
-        pass
-
-    @property
-    def vep_version(self):
-        """
-        Deprecated
-        """
-        return self._vep_version
-
-    @vep_version.setter
-    def vep_version(self, value):
-        """
-        Deprecated
-        """
-        pass
-
-    @property
-    def config_vep_version(self):
-        """
-        Deprecated
-        """
-        return self._config_vep_version
-
-    @config_vep_version.setter
-    def config_vep_version(self, value):
-        """
-        Deprecated
-        """
-        pass
-
-    @property
-    def config_loc(self):
-        """
-        Deprecated
-        """
-        return self._config_loc
-
-    @config_loc.setter
-    def config_loc(self, value):
-        """
-        Deprecated
         """
         pass
