+++ src/test/python/core/test_input_validator.py
@@ -3,7 +3,7 @@ import unittest
 import argparse
 from pathlib import Path
 from src.test.python.test_templates import teardown
-from src.main.python.resources.Validators import InputValidator
+from src.main.python.resources.validators import InputValidator
 from src.main.python.resources.utilities.utilities import get_project_root_dir
 from src.main.python.resources.processors.input_processor import InputProcessor
